# user-manage/api/user-manage/v1/user.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/user-manage/v1/user.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "UserService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/users": {
      "get": {
        "operationId": "UserService_ListUsers",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListUsersResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "usernamePrefix",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "nickNamePrefix",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "emailPrefix",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "status",
            "description": " - NORMAL: 正常\n - DISABLED: 禁用\n - LOCKED: 锁定，连续登录失败次数超过限制",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": [
              "UNKNOWN_STATUS",
              "NORMAL",
              "DISABLED",
              "LOCKED"
            ],
            "default": "UNKNOWN_STATUS"
          },
          {
            "name": "page.pageNo",
            "description": "page_no与page_token二选一",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "page.pageToken",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "page.pageSize",
            "description": "页面大小",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "page.orderBy",
            "description": "排序，符合SQL语法",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "UserService"
        ]
      },
      "post": {
        "operationId": "UserService_CreateUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1User"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateUserRequest"
            }
          }
        ],
        "tags": [
          "UserService"
        ]
      }
    },
    "/v1/users/{id}": {
      "get": {
        "operationId": "UserService_GetUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1User"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "UserService"
        ]
      },
      "delete": {
        "operationId": "UserService_DeleteUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "UserService"
        ]
      }
    },
    "/v1/users/{user.id}": {
      "put": {
        "operationId": "UserService_UpdateUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1User"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "user.id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "user": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "nickName": {
                      "type": "string"
                    },
                    "password": {
                      "type": "string"
                    },
                    "status": {
                      "$ref": "#/definitions/v1UserStatus"
                    },
                    "tel": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "createTime": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updateTime": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                },
                "updateMask": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "tags": [
          "UserService"
        ]
      },
      "patch": {
        "operationId": "UserService_UpdateUser2",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1User"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "user.id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "user": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "nickName": {
                      "type": "string"
                    },
                    "password": {
                      "type": "string"
                    },
                    "status": {
                      "$ref": "#/definitions/v1UserStatus"
                    },
                    "tel": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "createTime": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updateTime": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                },
                "updateMask": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "tags": [
          "UserService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1CreateUserRequest": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/v1User"
        }
      }
    },
    "v1ListUsersResponse": {
      "type": "object",
      "properties": {
        "users": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1User"
          }
        },
        "page": {
          "$ref": "#/definitions/v1PageResponse"
        }
      }
    },
    "v1PageRequest": {
      "type": "object",
      "properties": {
        "pageNo": {
          "type": "integer",
          "format": "int32",
          "title": "page_no与page_token二选一"
        },
        "pageToken": {
          "type": "string"
        },
        "pageSize": {
          "type": "integer",
          "format": "int32",
          "title": "页面大小"
        },
        "orderBy": {
          "type": "string",
          "title": "排序，符合SQL语法"
        }
      }
    },
    "v1PageResponse": {
      "type": "object",
      "properties": {
        "totalSize": {
          "type": "integer",
          "format": "int32"
        },
        "nextPageToken": {
          "type": "string"
        }
      }
    },
    "v1User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "nickName": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "status": {
          "$ref": "#/definitions/v1UserStatus"
        },
        "tel": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "createTime": {
          "type": "string",
          "format": "date-time"
        },
        "updateTime": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "v1UserStatus": {
      "type": "string",
      "enum": [
        "UNKNOWN_STATUS",
        "NORMAL",
        "DISABLED",
        "LOCKED"
      ],
      "default": "UNKNOWN_STATUS",
      "description": "- NORMAL: 正常\n - DISABLED: 禁用\n - LOCKED: 锁定，连续登录失败次数超过限制",
      "title": "用户状态"
    }
  }
}